move coach_profile to namespace
