+++ app/controllers/coach_profiles_controller.rb
@@ -1,8 +1,4 @@
 class CoachProfilesController < ApplicationController
-  def show
-    @user = current_user
-    @coach_profile = CoachProfile.find(params[:id])
-  end
 
   def new
     @coach_profile = CoachProfile.new
@@ -15,39 +11,10 @@ class CoachProfilesController < ApplicationController
     @coach_profile.user = @user
 
     if @coach_profile.save
-      redirect_to user_coach_profile_path(@user, @coach_profile)
+      redirect_to coach_coach_profile_path(@coach_profile)
     else
       render :new
     end
-
-    authorize @coach_profile
-  end
-
-  def edit
-    @coach_profile = CoachProfile.find(params[:id])
-    @user = current_user
-    @exists = true
-    authorize @coach_profile
-  end
-
-  def update
-    @coach_profile = CoachProfile.find(params[:id])
-    @coach_profile.update(coach_profile_params)
-    @user = current_user
-
-    if @coach_profile.save
-      redirect_to user_coach_profile_path(@user, @coach_profile)
-    else
-      render :new
-    end
-    authorize @coach_profile
-  end
-
-  def destroy
-    @user = current_user
-    @user.coach_profile.destroy
-    @user.save!
-    redirect_to user_path(@user)
 
     authorize @coach_profile
   end
