+++ app/policies/coach_profile_policy.rb
@@ -8,20 +8,4 @@ class CoachProfilePolicy < ApplicationPolicy
   def create?
     return true
   end
-
-  def edit?
-    update?
-  end
-
-  def update?
-    record.user == user
-  end
-
-  def show?
-    true
-  end
-
-  def destroy?
-    record.user == user
-  end
 end
