+++ app/views/shared/_navbar.html.erb
@@ -23,7 +23,7 @@
             <%= link_to "Coach Lessons", coach_lessons_path, class: "nav-link coach-lessons-btn #{"hide" if !coach_path? } #{"active" if current_page?(coach_lessons_path) || current_page?(upcoming_coach_lessons_path)|| current_page?(history_coach_lessons_path)}" %>
           </li>
           <li class="nav-item">
-            <%= link_to "Coach Profile", user_coach_profile_path(current_user, current_user.coach_profile), class: "nav-link coach-profile-btn #{"hide" if !coach_path? } #{"active" if current_page?(user_coach_profile_path(current_user, current_user.coach_profile))}" %>
+            <%= link_to "Coach Profile", coach_coach_profile_path(current_user.coach_profile), class: "nav-link coach-profile-btn #{"hide" if !coach_path? } #{"active" if current_page?(coach_coach_profile_path(current_user.coach_profile))}" %>
           </li>
           <li class="nav-item">
             <div class="custom-control custom-switch">
