+++ config/routes.rb
@@ -2,11 +2,6 @@ Rails.application.routes.draw do
   devise_for :users
   root to: 'pages#home'
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
-
-  resources :users, only: :show do
-    resources :coach_profiles, only: [:show, :new, :create, :edit, :update, :destroy]
-  end
-
   namespace :coach do
     resources :lessons do
       collection do
@@ -14,7 +9,14 @@ Rails.application.routes.draw do
         get :history
       end
     end
+
+    resources :coach_profiles, only: [:show, :edit, :update, :destroy]
   end
+
+  resources :users, only: :show do
+    resources :coach_profiles, only: [:new, :create]
+  end
+
 
   resources :sports, only: [:index] do
     resources :coaches, only: [:index]
